Move JAX sort dispatch to its own module
