+++ tests/link/jax/test_tensor_basic.py
@@ -218,15 +218,6 @@ def test_tri():
     compare_jax_and_py(fgraph, [])
 
 
-@pytest.mark.parametrize("axis", [None, -1])
-def test_sort(axis):
-    x = matrix("x", shape=(2, 2), dtype="float64")
-    out = pytensor.tensor.sort(x, axis=axis)
-    fgraph = FunctionGraph([x], [out])
-    arr = np.array([[1.0, 4.0], [5.0, 2.0]])
-    compare_jax_and_py(fgraph, [arr])
-
-
 def test_tri_nonconcrete():
     """JAX cannot JIT-compile `jax.numpy.tri` when arguments are not concrete values."""
